Fix compiler warning

+++ Bugsnag/Payload/BugsnagStackframe.m
@@ -103,7 +103,7 @@ BugsnagStackframeType const BugsnagStackframeTypeCocoa = @"cocoa";
 }
 
 + (NSArray<BugsnagStackframe *> *)stackframesWithCallStackReturnAddresses:(NSArray<NSNumber *> *)callStackReturnAddresses {
-    int length = callStackReturnAddresses.count;
+    int length = (int)callStackReturnAddresses.count;
     uintptr_t addresses[length];
     for (int i = 0; i < length; i++) {
         addresses[i] = (uintptr_t)callStackReturnAddresses[i].unsignedLongLongValue;
